nmod_poly_factor print
